+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Braintree Android SDK Release Notes
 
-## unreleased
+## Unreleased
+* Adds `invalidateClientToken` method to `BraintreeClient`, which invalidates the cached client
+token authorization provided by the `ClientTokenProvider`
 * Add `isExplicitCancelation` parameter to `UserCanceledException`
 
 ## 4.10.1